+++ src/ui/lang/es.rs
@@ -62,6 +62,7 @@ pub fn text(msg: Msg) -> Cow<'static, str> {
         Msg::SpeakerToggleOn => "\u{1F50A} Activado".into(),
         Msg::SpeakerToggleOff => "\u{1F50A} Altavoz".into(),
         Msg::SpeakerPlain => "Altavoz".into(),
+        Msg::EarpiecePlain => "Auricular".into(),
         Msg::AddHandle => "+ Handle".into(),
         Msg::AddHandlePlain => "Agregar handle".into(),
         Msg::BackToContact => "\u{2039} Contacto".into(),