+++ src/ui/lang/mi.rs
@@ -62,6 +62,7 @@ pub fn text(msg: Msg) -> Cow<'static, str> {
         Msg::SpeakerToggleOn => "\u{1F50A} Kā ana".into(),
         Msg::SpeakerToggleOff => "\u{1F50A} Pūoro".into(),
         Msg::SpeakerPlain => "Pūoro".into(),
+        Msg::EarpiecePlain => "Pūoro taringa".into(),
         Msg::AddHandle => "+ Handle".into(),
         Msg::AddHandlePlain => "Tāpiri handle".into(),
         Msg::BackToContact => "\u{2039} Hoa".into(),